Change type of socket_hint_type from int to unsigned short.

+++ include/network/socket-hint-type.h
@@ -18,7 +18,7 @@
 
 namespace Network
 {
-    using socket_hint_type = int;
+    using socket_hint_type = unsigned short;
 }
 
 #endif
